Create a factor for the ActorStack
This enables Sufia to customize the ActorStack slightly.

Fixes projecthydra/sufia#1776

Will be used to complete work on:
Ref projecthydra/sufia#1699
