+++ app/services/curation_concerns/curation_concern.rb
@@ -1,46 +1,15 @@
 module CurationConcerns
-  module CurationConcern
-    # Returns the top-level actor on the stack
+  class CurationConcern
+    class_attribute :actor_factory
+    self.actor_factory = CurationConcerns::ActorFactory
+
+    # A consumer of this method can inject a different factory
+    # into this class in order to change the behavior of this method.
+    # @param [ActiveFedora::Base] curation_concern a work to be updated
+    # @param [User] current_user the depositor/updater of the work
+    # @return [#create, #update] an actor that can create and update the work
     def self.actor(curation_concern, current_user)
-      ActorStack.new(curation_concern, current_user,
-                     [AddToCollectionActor,
-                      AssignRepresentativeActor,
-                      AttachFilesActor,
-                      ApplyOrderActor,
-                      InterpretVisibilityActor,
-                      model_actor(curation_concern),
-                      AssignIdentifierActor])
-    end
-
-    def self.model_actor(curation_concern)
-      actor_identifier = curation_concern.class.to_s.split('::').last
-      "CurationConcerns::#{actor_identifier}Actor".constantize
-    end
-
-    class ActorStack
-      attr_reader :curation_concern, :user, :first_actor_class, :more_actors
-      def initialize(curation_concern, user, more_actors)
-        @curation_concern = curation_concern
-        @user = user
-        @more_actors = more_actors
-        @first_actor_class = @more_actors.shift || RootActor
-      end
-
-      def inner_stack
-        ActorStack.new(curation_concern, user, more_actors)
-      end
-
-      def actor
-        first_actor_class.new(curation_concern, user, inner_stack)
-      end
-
-      def create(attributes)
-        actor.create(attributes.with_indifferent_access)
-      end
-
-      def update(attributes)
-        actor.update(attributes.with_indifferent_access)
-      end
+      actor_factory.build(curation_concern, current_user)
     end
   end
 end
